+++ microservices/BookingService/BookingService.Business/Services/BookingBusinessService.cs
@@ -9,6 +9,10 @@ using EmployeeService.ClientHttp.Interfaces;
 
 namespace BookingService.Business.Services
 {
+    /// <summary>
+    /// Implementazione del servizio di business per la gestione delle prenotazioni.
+    /// Coordina repository, servizi esterni e pubblicazione eventi.
+    /// </summary>
     public class BookingBusinessService : IBookingBusinessService
     {
         private readonly IBookingRepository _repository;
@@ -16,6 +20,9 @@ namespace BookingService.Business.Services
         private readonly IEmployeeClient _employeeClient;
         private readonly IBookingEventProducer _eventProducer;
 
+        /// <summary>
+        /// Inizializza una nuova istanza di <see cref="BookingBusinessService"/>.
+        /// </summary>
         public BookingBusinessService(
             IBookingRepository repository,
             ICarClient carClient,
@@ -28,20 +35,31 @@ namespace BookingService.Business.Services
             _eventProducer = eventProducer;
         }
 
-        public async Task<int> CreateBookingAsync(CreateBookingDto dto)
+        /// <summary>
+        /// Crea una nuova prenotazione dopo aver validato impiegato e veicolo.
+        /// </summary>
+        /// <param name="dto">Dati necessari alla creazione della prenotazione.</param>
+        /// <returns>Identificativo della prenotazione creata.</returns>
+        /// <exception cref="InvalidOperationException">
+        /// Viene lanciata se l’impiegato non è idoneo o l’auto non è disponibile.
+        /// </exception>
+        public async Task<short> CreateBookingAsync(CreateBookingDto dto)
         {
-            // 1. Validazione Impiegato (Sincrona) 
+            // 1. Validazione impiegato
             var isEligible = await _employeeClient.GetEmployeeEligibilityAsync(dto.EmployeeId);
-            if (!isEligible) throw new InvalidOperationException("L'impiegato non è abilitato alla guida.");
+            if (!isEligible)
+                throw new InvalidOperationException("L'impiegato non è abilitato alla guida.");
 
             var alreadyHasBooking = await _repository.HasActiveBookingAsync(dto.EmployeeId);
-            if (alreadyHasBooking) throw new InvalidOperationException("L'impiegato ha già una prenotazione attiva.");
+            if (alreadyHasBooking)
+                throw new InvalidOperationException("L'impiegato ha già una prenotazione attiva.");
 
-            // 2. Validazione Auto (Sincrona) 
+            // 2. Validazione auto
             var car = await _carClient.GetCarByIdAsync(dto.CarId);
-            if (car == null || car.Status != "AVAILABLE") throw new InvalidOperationException("L'auto non è disponibile.");
+            if (car == null || car.Status != "AVAILABLE")
+                throw new InvalidOperationException("L'auto non è disponibile.");
 
-            // 3. Creazione record locale
+            // 3. Creazione prenotazione locale
             var booking = new Booking
             {
                 EmployeeId = dto.EmployeeId,
@@ -54,22 +72,28 @@ namespace BookingService.Business.Services
             await _repository.AddAsync(booking);
             await _repository.SaveChangesAsync();
 
-            // 4. Side effect: Aggiorna stato auto a IN_USE 
+            // 4. Side effect: aggiornamento stato veicolo
             await _carClient.UpdateCarStatusAsync(dto.CarId, "IN_USE");
 
             return booking.Id;
         }
 
+        /// <summary>
+        /// Completa una prenotazione e pubblica l’evento di viaggio completato.
+        /// </summary>
+        /// <param name="bookingId">Identificativo della prenotazione.</param>
+        /// <param name="kmTraveled">Chilometri percorsi.</param>
         public async Task ProcessCompletionAsync(short bookingId, double kmTraveled)
         {
             var booking = await _repository.GetByIdAsync(bookingId);
-            if (booking == null || booking.Status != BookingStatus.ACTIVE) return;
+            if (booking == null || booking.Status != BookingStatus.ACTIVE)
+                return;
 
-            // Aggiorna stato locale
+            // Aggiornamento stato locale
             booking.Status = BookingStatus.COMPLETED;
             await _repository.SaveChangesAsync();
 
-            // Pubblica evento per il Car Service 
+            // Pubblicazione evento
             await _eventProducer.PublishTripCompletedAsync(new TripCompletedEvent
             {
                 VehicleId = booking.CarId,
@@ -79,16 +103,39 @@ namespace BookingService.Business.Services
             });
         }
 
+        /// <summary>
+        /// Recupera tutte le prenotazioni presenti nel sistema.
+        /// </summary>
         public async Task<IEnumerable<BookingDto>> GetAllBookingsAsync()
         {
             var list = await _repository.GetAllAsync();
-            return list.Select(b => new BookingDto(b.Id, b.EmployeeId, b.CarId, b.StartTime, b.EndTime, b.Status.ToString()));
+
+            return list.Select(b =>
+                new BookingDto(
+                    b.Id,
+                    b.EmployeeId,
+                    b.CarId,
+                    b.StartTime,
+                    b.EndTime,
+                    b.Status.ToString()));
         }
 
+        /// <summary>
+        /// Recupera una prenotazione tramite ID.
+        /// </summary>
         public async Task<BookingDto?> GetBookingByIdAsync(short id)
         {
             var b = await _repository.GetByIdAsync(id);
-            return b == null ? null : new BookingDto(b.Id, b.EmployeeId, b.CarId, b.StartTime, b.EndTime, b.Status.ToString());
+
+            return b == null
+                ? null
+                : new BookingDto(
+                    b.Id,
+                    b.EmployeeId,
+                    b.CarId,
+                    b.StartTime,
+                    b.EndTime,
+                    b.Status.ToString());
         }
     }
 }
